adapt to WSL

+++ LUNA16Challege/ResNet3d/implementation/Resnet3d.py
@@ -253,8 +253,8 @@ class RestNet3dModule(object):
                                         (self.image_depth, self.image_height, self.image_width, self.channels))
           batch_xs[num, :, :, :] = batchimage
          # Extracting images and labels from given data
-         batch_xs = batch_xs.astype(np.float)
-         batch_ys = batch_ys.astype(np.float)
+         batch_xs = batch_xs.astype(np.float32)
+         batch_ys = batch_ys.astype(np.float32)
          # Normalize from [0:255] => [0.0:1.0]
          batch_xs = np.multiply(batch_xs, 1.0 / 255.0)
          batch_xs=np.float32(batch_xs) 
@@ -279,7 +279,7 @@ class RestNet3dModule(object):
         self.phase = 1
         test_images = np.reshape(test_images, (
             test_images.shape[0], test_images.shape[1], test_images.shape[2], test_images.shape[3], 1))
-        test_images = test_images.astype(np.float)
+        test_images = test_images.astype(np.float32)
         test_images = np.multiply(test_images, 1.0 / 255.0)
         test_images=np.float32(test_images)
